refactor: [WIP] update architecture
