+++ src/obsidian/nlt-import-modal.tsx
@@ -1,6 +1,6 @@
 import { App, Modal } from "obsidian";
 import { NLTView } from "./nlt-view";
-import { TableState } from "../data/types";
+import { TableState } from "../shared/table-state/types";
 import { Root, createRoot } from "react-dom/client";
 import ImportApp from "../react/import-app";
 import {
