+++ src/react/table-app/date-cell-edit/components/DateFormatMenu/index.tsx
@@ -1,8 +1,7 @@
 import Menu from "src/react/shared/menu";
 import MenuItem from "src/react/shared/menu-item";
-import { DateFormat } from "src/data/types";
-import { getDisplayNameForDateFormat } from "src/shared/table-state/utils";
-
+import { getDisplayNameForDateFormat } from "src/shared/table-state/display-name";
+import { DateFormat } from "src/shared/table-state/types";
 interface Props {
 	id: string;
 	top: number;
