+++ src/react/table-app/function-cell/numberFunction.ts
@@ -7,7 +7,11 @@ import {
 	getSum,
 } from "./arithmetic";
 import { stringToCurrencyString } from "src/shared/conversion";
-import { CellType, CurrencyType, NumberFunction } from "src/data/types";
+import {
+	CellType,
+	CurrencyType,
+	NumberFunction,
+} from "src/shared/table-state/types";
 import { round2Digits } from "./utils";
 
 export const getNumberFunctionContent = (
