+++ src/react/table-app/header-cell/components/HeaderMenu/components/BaseMenu/index.tsx
@@ -2,11 +2,11 @@ import Divider from "src/react/shared/divider";
 import MenuItem from "src/react/shared/menu-item";
 import Padding from "src/react/shared/padding";
 import Stack from "src/react/shared/stack";
-import { CellType, SortDir } from "src/data/types";
-import { getDisplayNameForCellType } from "src/shared/table-state/utils";
+import { CellType, SortDir } from "src/shared/table-state/types";
 import { SubmenuType } from "../../types";
 import { IconType } from "src/react/shared/icon/types";
 import { useFocusMenuInput } from "src/shared/hooks";
+import { getDisplayNameForCellType } from "src/shared/table-state/display-name";
 
 interface Props {
 	isMenuVisible: boolean;
