+++ src/react/table-app/header-cell/components/HeaderMenu/components/CurrencySubmenu/index.tsx
@@ -1,7 +1,7 @@
 import MenuItem from "src/react/shared/menu-item";
 import Submenu from "../Submenu";
-import { CurrencyType } from "src/data/types";
-import { getDisplayNameForCurrencyType } from "src/shared/table-state/utils";
+import { CurrencyType } from "src/shared/table-state/types";
+import { getDisplayNameForCurrencyType } from "src/shared/table-state/display-name";
 
 interface Props {
 	title: string;
