+++ src/react/table-app/header-cell/components/HeaderMenu/components/DateFormatSubmenu/index.tsx
@@ -1,7 +1,7 @@
 import MenuItem from "src/react/shared/menu-item";
 import Submenu from "../Submenu";
-import { DateFormat } from "src/data/types";
-import { getDisplayNameForDateFormat } from "src/shared/table-state/utils";
+import { DateFormat } from "src/shared/table-state/types";
+import { getDisplayNameForDateFormat } from "src/shared/table-state/display-name";
 
 interface Props {
 	title: string;
