+++ src/react/table-app/header-cell/components/HeaderMenu/components/OptionSubmenu/components/CurrencyMenu/index.tsx
@@ -1,7 +1,7 @@
 import Menu from "src/react/shared/menu";
 import MenuItem from "src/react/shared/menu-item";
-import { CurrencyType } from "src/data/types";
-import { getDisplayNameForCurrencyType } from "src/shared/table-state/utils";
+import { getDisplayNameForCurrencyType } from "src/shared/table-state/display-name";
+import { CurrencyType } from "src/shared/table-state/types";
 
 interface Props {
 	id: string;
