+++ src/react/table-app/last-edited-time-cell/index.tsx
@@ -1,6 +1,6 @@
 import { unixTimeToDateTimeString } from "src/shared/date/date-conversion";
 import { useOverflowClassName } from "src/shared/spacing/hooks";
-import { DateFormat } from "src/data/types";
+import { DateFormat } from "src/shared/table-state/types";
 
 interface Props {
 	value: number;
