+++ src/react/table-app/option-bar/filter/filter-menu.tsx
@@ -6,8 +6,8 @@ import Icon from "src/react/shared/icon";
 import { IconType } from "src/react/shared/icon/types";
 import { Button } from "src/react/shared/button";
 import FilterRow from "./filter-row";
-import { FilterRule, FilterType, Tag } from "src/data/types";
-import { ColumnIdError } from "src/shared/table-state/error";
+import { FilterRule, FilterType, Tag } from "src/shared/table-state/types";
+import { ColumnIdError } from "src/shared/table-state/table-error";
 import Text from "src/react/shared/text";
 
 interface Props {
