+++ src/react/table-app/table/table-body-row.tsx
@@ -1,6 +1,6 @@
 import { useTableState } from "src/shared/table-state/table-state-context";
 import { TableDataTransferItem } from "./types";
-import { SortDir } from "src/data/types";
+import { SortDir } from "src/shared/table-state/types";
 
 interface TableBodyRowProps {
 	children?: React.ReactNode;
