+++ src/react/table-app/tag-cell-edit/menu-body.tsx
@@ -1,7 +1,7 @@
 import Padding from "src/react/shared/padding";
 import Text from "src/react/shared/text";
 import { Color } from "src/shared/types";
-import { Tag } from "src/data/types";
+import { Tag } from "src/shared/table-state/types";
 import CreateTag from "./create-tag";
 import SelectableTag from "./selectable-tag";
