+++ src/shared/table-state/filter-by-rules.ts
@@ -7,8 +7,8 @@ import {
 	FilterType,
 	TableState,
 	Tag,
-} from "src/data/types";
-import { ColumnIdError } from "../table-state/error";
+} from "src/shared/table-state/types";
+import { ColumnIdError } from "./table-error";
 
 export const filterBodyRowsByRules = (prevState: TableState): BodyRow[] => {
 	const { columns, bodyCells, bodyRows, filterRules, tags } = prevState.model;
