+++ src/shared/table-state/use-filter-rules.ts
@@ -1,7 +1,7 @@
 import { SetStateAction } from "react";
-import { FilterType, TableState } from "src/data/types";
+import { FilterType, TableState } from "src/shared/table-state/types";
 import { useLogger } from "../logger";
-import { addRule, deleteRule, updateRule } from "./state-operations";
+import { addRule, deleteRule, updateRule } from "./filter-state-operations";
 import { filterBodyRowsByRules } from "./filter-by-rules";
 
 export const useFilterRules = (
