+++ src/shared/validators.ts
@@ -1,3 +1,4 @@
+import { GeneralFunction, NumberFunction } from "src/shared/table-state/types";
 import {
 	NUMBER_REGEX,
 	TAG_REGEX,
@@ -24,4 +25,16 @@ export const isCheckboxChecked = (input: string): boolean => {
 
 export const isTag = (input: string): boolean => {
 	return input.match(TAG_REGEX) !== null;
+};
+
+export const isGeneralFunction = (
+	value: GeneralFunction | NumberFunction
+): value is GeneralFunction => {
+	return Object.values(GeneralFunction).includes(value as GeneralFunction);
+};
+
+export const isNumberFunction = (
+	value: GeneralFunction | NumberFunction
+): value is NumberFunction => {
+	return Object.values(NumberFunction).includes(value as NumberFunction);
 };
